// src/supabase.d.ts
export type Database = {
// ...
  public: {
    Tables: {
      points: {
        Row: {
          created_at: string
          id: number
          location: unknown
        }
        Insert: {
          created_at?: string
          id?: number
          location: unknown
        }
        Update: {
          created_at?: string
          id?: number
          location?: unknown
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}